+++ lib/src/lz4/ffi/library.dart
@@ -4,7 +4,7 @@
 
 import 'dart:ffi';
 
-import '../../framework/native/open_library.dart';
+import '../../framework/native/library/open_library.dart';
 import 'constants.dart';
 import 'functions.dart';
 import 'types.dart';